+++ test/unit/fuzz-tools.test.js
@@ -321,6 +321,7 @@ import * as masonRadialHeadV341 from '../../lib/mason-radial-head-v341.js';
 import * as hawkinsTalarV342 from '../../lib/hawkins-talar-v342.js';
 import * as sandersCalcanealV343 from '../../lib/sanders-calcaneal-v343.js';
 import * as ficatArletV344 from '../../lib/ficat-arlet-v344.js';
+import * as lichtmanKienbockV345 from '../../lib/lichtman-kienbock-v345.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -587,6 +588,7 @@ const MODULES = {
   'hawkins-talar-v342.js': hawkinsTalarV342,
   'sanders-calcaneal-v343.js': sandersCalcanealV343,
   'ficat-arlet-v344.js': ficatArletV344,
+  'lichtman-kienbock-v345.js': lichtmanKienbockV345,
   'tb-testing.js': tbTesting,
 };